refactor: moved config to its own file
